Run commands button
Added button to toggle whether commands from the queue are being run. Also changed the button for accepting commands to match the new style, which is much cleaner.

Other change: Changed the evaluation of programs in the database from using ```eval()``` to using``` json.loads()```, so that nothing is executed when the information is evaluated. Better for security